## Summary
- fix automatic verb registration system and wire it up to the test harness
- complete 64-bit alignment work for all numeric types, math operations, and datetime values
- align headless notifydialog with GUI semantics by printing to stdout and blocking for Enter
- clean remaining compiler warnings (sign-compare, pointer-sign, unused params) and update numeric/date checklist docs
- remove accidentally tracked test binaries, expand .gitignore for test executables, keep Paige binaries untouched

## So What
- headless UX now mirrors blocking dialog behavior without GUI dependencies
- warning-free builds keep the 64-bit BE64 work stable and easier to maintain
- repository no longer carries transient test artifacts, reducing churn for future work

## Testing
- make -C tests test  # file_verb_tests still skipped because file.open remains unimplemented in headless


<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> Upgrade runtime and serialization to 64‑bit (BE64) with modern packers, implement key headless verbs, stabilize CLI eval/migration, and add targeted tests and tooling.
> 
> - **Core/runtime (64‑bit)**:
>   - Widen `tyvaluedata` (`int/long/date`) to 64‑bit; update setters/coercions, arithmetic/bitwise ops, and constants (`longinfinity`).
>   - Switch time APIs and string conversions to 64‑bit; rework `numbertostring`/`stringtonumber`, `hexstringtonumber`/`numbertohexstring`.
> - **Serialization (v7 BE64)**:
>   - Add modern BE64 value payloads and pack/unpack (`tydisksymbolrecord_v7`); split legacy vs modern visitors/dispatch.
>   - Modernize outline/menu packers (memcpy for packed 64‑bit fields); add list unpack dispatch and headless tracing.
>   - Migration/format mode fixes: reject out‑of‑range versions, force legacy during migrate, seed mode from header to avoid remigrate.
> - **Headless/CLI**:
>   - Implement `clock.now/ticks/milliseconds`, `math.random`, and `string.length/upper/lower`; headless `notifydialog` blocks on Enter; resource loader stubbed for headless.
>   - CLI: execute long scripts via `langrunhandle`; ensure `frontier-cli` builds for tests.
> - **Tests/Tooling**:
>   - Add `hashpack_modern_tests`, `langvalue_64_tests`, strengthen `save_migration_tests`; introduce `tools/run_headless_tests.sh`.
>   - Expand test Makefiles; skip `file_verb_tests` until `file.open` exists; broaden `.gitignore`.
> - **Docs**: Update AGENTS/CLAUDE and planning notes for headless flow and 64‑bit plan.
> 
> <sup>Written by [Cursor Bugbot](https://cursor.com/dashboard?tab=bugbot) for commit b91f9433ad2bbdc318218eee475d32e0d574ad15. This will update automatically on new commits. Configure [here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->

+++ Common/source/opverbs.c
@@ -593,7 +593,7 @@ static boolean opverbinmemory (hdloutlinevariable hv) {
 		short versionnumber;
 		boolean islegacy = false;
 
-		if (gethandlesize(hpackedoutline) >= sizeof(short)) {
+		if (gethandlesize(hpackedoutline) >= (long) sizeof(short)) {
 			versionnumber = *(short *)(*hpackedoutline);
 			disktomemshort(versionnumber);
